refactor(user): modularize routes with controllers, services, and moved getDocumentTypeLabel to utils ffor cleaner structure
